03-declarative-agent-manifest/stub-endpoint/requirements.txt
flask>=3.0.0
azure-ai-projects>=1.0.0b5
azure-identity>=1.15.0
